pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "generativeai>=0.0.1",
    "google>=3.0.0",
    "python-dotenv>=1.1.0",
]
